+++ window/backend/renderer_sdl/render.go
@@ -116,7 +116,7 @@ func (sr *sdlRender) createRendererTexture() *sdl.Texture {
 		log.Printf("ERROR: %v", err)
 		return nil
 	}
-	texture, err := sr.renderer.CreateTexture(uint32(sdl.PIXELFORMAT_RGBA32), sdl.TEXTUREACCESS_TARGET, w, h)
+	texture, err := sr.renderer.CreateTexture(sdl.PIXELFORMAT_RGBA32, sdl.TEXTUREACCESS_TARGET, w, h)
 	if err != nil {
 		log.Printf("ERROR: %v", err)
 		return nil
